Fixing pill navigation, like-boxes and logo
Much was changed, index.html is now starting to look really incomprehensive!

The navigation bar for each page now works with corresponding nav-panes.
The bad thing with this concept is that the code is now very hard to
read and not so interchangeable or automatic. The nice thing is that it
has the correct functionality. Oh, and it's also not possible to
reference the current page with a link, and a solution with scripts
looking up the URI is not optimal. Look into this!

The html5 version of facebook like boxes is now implemented and works
when read from a server. Might look a bit odd on xs, but I don't think
that's a problem. The boxes are monitored via facebook insights.

Logos are now in place and change according to page. The bad thing is
that they are not hidden/changed on smaller devices. The good thing is
that it looks great!